editing: extract ScriptEditService from ObjectCommandController
Continues the controller split: all map-script editing — attach/detach/spatial
scripts plus the script-id/object-id allocation and section snapshot/restore
machinery — moves into ScriptEditService, built on the shared UndoBatcher.
snapshotScripts/restoreScripts/eraseScript are public so clearElevationObjects can
still prune the scripts of the objects it removes. ObjectCommandController keeps the
public script methods as thin delegations and drops the now-unused
MapScript/array includes. 235 tests pass.
