+++ src/ui/editing/ObjectCommandController.cpp
@@ -37,6 +37,7 @@ ObjectCommandController::ObjectCommandController(resource::GameResources& resour
     , _batcher(undoStack, std::move(onStackChanged))
     , _tileService(map, _batcher, std::move(ensureElevationTiles), std::move(getCurrentElevation), std::move(updateTileSprite))
     , _inventoryService(_batcher)
+    , _scriptService(map, _batcher)
     , _refreshObjects(std::move(refreshObjects))
     , _reloadTiles(std::move(reloadTiles)) {
 }
@@ -520,25 +521,25 @@ bool ObjectCommandController::clearElevationObjects(int elevation) {
     // Removing the objects must also remove their attached scripts, otherwise the
     // map keeps orphaned MapScript records pointing at OIDs that no longer exist.
     // Prune by SID so the objects' own map_scripts_pid stays intact for undo.
-    ScriptSections beforeScripts = snapshotScripts();
+    ScriptEditService::ScriptSections beforeScripts = _scriptService.snapshotScripts();
     it->second.clear();
     for (const auto& obj : before) {
         if (obj && obj->map_scripts_pid != -1) {
-            eraseScript(static_cast<uint32_t>(obj->map_scripts_pid));
+            _scriptService.eraseScript(static_cast<uint32_t>(obj->map_scripts_pid));
         }
     }
-    ScriptSections afterScripts = snapshotScripts();
+    ScriptEditService::ScriptSections afterScripts = _scriptService.snapshotScripts();
 
     UndoCommand cmd;
     cmd.description = "Clear Elevation Objects";
     cmd.undo = [this, elevation, before, beforeScripts]() {
         _map->getMapFile().map_objects[elevation] = before;
-        restoreScripts(beforeScripts);
+        _scriptService.restoreScripts(beforeScripts);
         _refreshObjects();
     };
     cmd.redo = [this, elevation, after, afterScripts]() {
         _map->getMapFile().map_objects[elevation] = after;
-        restoreScripts(afterScripts);
+        _scriptService.restoreScripts(afterScripts);
         _refreshObjects();
     };
     // The edit was already applied above; only register the undo/redo handlers.
@@ -616,174 +617,16 @@ bool ObjectCommandController::copyElevation(int fromElevation, int toElevation) 
     return pushCommand(std::move(cmd));
 }
 
-uint32_t ObjectCommandController::allocateScriptId(int section) const {
-    if (!_map || section < 0 || section >= Map::SCRIPT_SECTIONS) {
-        return 0;
-    }
-    uint32_t nextId = 0;
-    for (const auto& s : _map->getMapFile().map_scripts[section]) {
-        nextId = std::max(nextId, MapScript::sidIndex(s.pid) + 1);
-    }
-    return nextId;
-}
-
-uint32_t ObjectCommandController::allocateObjectId() const {
-    if (!_map) {
-        return 1;
-    }
-    uint32_t maxId = 0;
-    const auto& mapFile = _map->getMapFile();
-    for (const auto& [elevation, objects] : mapFile.map_objects) {
-        for (const auto& obj : objects) {
-            if (obj) {
-                maxId = std::max(maxId, obj->unknown0);
-            }
-        }
-    }
-    for (int section = 0; section < Map::SCRIPT_SECTIONS; ++section) {
-        for (const auto& s : mapFile.map_scripts[section]) {
-            maxId = std::max(maxId, s.script_oid);
-        }
-    }
-    return maxId + 1;
-}
-
-ObjectCommandController::ScriptSections ObjectCommandController::snapshotScripts() const {
-    static_assert(SCRIPT_SECTIONS == Map::SCRIPT_SECTIONS,
-        "ScriptSections snapshot size must match Map::SCRIPT_SECTIONS");
-    ScriptSections snapshot;
-    const auto& mapFile = _map->getMapFile();
-    for (int section = 0; section < Map::SCRIPT_SECTIONS; ++section) {
-        snapshot.sections[section] = mapFile.map_scripts[section];
-        snapshot.counts[section] = mapFile.scripts_in_section[section];
-    }
-    return snapshot;
-}
-
-void ObjectCommandController::restoreScripts(const ScriptSections& snapshot) {
-    auto& mapFile = _map->getMapFile();
-    for (int section = 0; section < Map::SCRIPT_SECTIONS; ++section) {
-        mapFile.map_scripts[section] = snapshot.sections[section];
-        mapFile.scripts_in_section[section] = snapshot.counts[section];
-    }
-}
-
-void ObjectCommandController::eraseScript(uint32_t sid) {
-    const int section = MapScript::sidSection(sid);
-    if (section < 0 || section >= Map::SCRIPT_SECTIONS) {
-        return;
-    }
-    auto& vec = _map->getMapFile().map_scripts[section];
-    std::erase_if(vec, [sid](const MapScript& s) { return s.pid == sid; });
-    _map->getMapFile().scripts_in_section[section] = static_cast<int>(vec.size());
-}
-
-void ObjectCommandController::removeObjectScript(MapObject& object) {
-    if (!_map || object.map_scripts_pid == -1) {
-        return;
-    }
-    eraseScript(static_cast<uint32_t>(object.map_scripts_pid));
-    object.map_scripts_pid = -1;
-}
-
-void ObjectCommandController::applyScriptSnapshot(int section, const std::shared_ptr<MapObject>& object,
-    const std::vector<MapScript>& sectionScripts, uint32_t oid, int32_t sid) {
-    if (!_map || section < 0 || section >= Map::SCRIPT_SECTIONS) {
-        return;
-    }
-    auto& mapFile = _map->getMapFile();
-    mapFile.map_scripts[section] = sectionScripts;
-    mapFile.scripts_in_section[section] = static_cast<int>(sectionScripts.size());
-    if (object) {
-        object->unknown0 = oid;
-        object->map_scripts_pid = sid;
-    }
-}
-
-bool ObjectCommandController::recordScriptEdit(const std::string& description, int section,
-    const std::shared_ptr<MapObject>& object,
-    std::vector<MapScript> beforeSection, uint32_t beforeOid, int32_t beforeSid) {
-    if (!_map) {
-        return false;
-    }
-    // The caller already applied the edit; capture the resulting "after" state.
-    std::vector<MapScript> afterSection = _map->getMapFile().map_scripts[section];
-    uint32_t afterOid = object ? object->unknown0 : 0;
-    int32_t afterSid = object ? object->map_scripts_pid : -1;
-
-    UndoCommand cmd;
-    cmd.description = description;
-    cmd.undo = [this, section, object, beforeSection = std::move(beforeSection), beforeOid, beforeSid]() {
-        applyScriptSnapshot(section, object, beforeSection, beforeOid, beforeSid);
-    };
-    cmd.redo = [this, section, object, afterSection = std::move(afterSection), afterOid, afterSid]() {
-        applyScriptSnapshot(section, object, afterSection, afterOid, afterSid);
-    };
-    return pushCommand(std::move(cmd));
-}
-
-bool ObjectCommandController::attachScript(const std::shared_ptr<MapObject>& object,
-    int scriptType, uint32_t programIndex) {
-    if (!_map || !object || scriptType < 0 || scriptType >= Map::SCRIPT_SECTIONS) {
-        return false;
-    }
-    const int section = scriptType;
-    auto& mapFile = _map->getMapFile();
-
-    std::vector<MapScript> beforeSection = mapFile.map_scripts[section];
-    const uint32_t beforeOid = object->unknown0;
-    const int32_t beforeSid = object->map_scripts_pid;
-
-    // Replace any existing script (same object type -> same section).
-    removeObjectScript(*object);
-
-    const uint32_t scriptId = allocateScriptId(section);
-    const uint32_t oid = allocateObjectId();
-    MapScript script = MapScript::makeObjectScript(static_cast<MapScript::ScriptType>(scriptType),
-        scriptId, programIndex, oid);
-    mapFile.map_scripts[section].push_back(script);
-    mapFile.scripts_in_section[section] = static_cast<int>(mapFile.map_scripts[section].size());
-    object->unknown0 = oid;
-    object->map_scripts_pid = static_cast<int32_t>(script.pid);
-
-    return recordScriptEdit("Attach Script", section, object, std::move(beforeSection), beforeOid, beforeSid);
+bool ObjectCommandController::attachScript(const std::shared_ptr<MapObject>& object, int scriptType, uint32_t programIndex) {
+    return _scriptService.attachScript(object, scriptType, programIndex);
 }
 
 bool ObjectCommandController::detachScript(const std::shared_ptr<MapObject>& object) {
-    if (!_map || !object || object->map_scripts_pid == -1) {
-        return false;
-    }
-    const int section = MapScript::sidSection(static_cast<uint32_t>(object->map_scripts_pid));
-    if (section < 0 || section >= Map::SCRIPT_SECTIONS) {
-        return false;
-    }
-
-    std::vector<MapScript> beforeSection = _map->getMapFile().map_scripts[section];
-    const uint32_t beforeOid = object->unknown0;
-    const int32_t beforeSid = object->map_scripts_pid;
-
-    removeObjectScript(*object);
-
-    return recordScriptEdit("Detach Script", section, object, std::move(beforeSection), beforeOid, beforeSid);
+    return _scriptService.detachScript(object);
 }
 
 bool ObjectCommandController::addSpatialScript(uint32_t programIndex, int tile, int elevation, int radius) {
-    if (!_map) {
-        return false;
-    }
-    const int section = static_cast<int>(MapScript::ScriptType::SPATIAL);
-    auto& mapFile = _map->getMapFile();
-
-    std::vector<MapScript> beforeSection = mapFile.map_scripts[section];
-
-    const uint32_t scriptId = allocateScriptId(section);
-    const uint32_t oid = allocateObjectId();
-    MapScript script = MapScript::makeSpatialScript(scriptId, programIndex,
-        static_cast<uint32_t>(tile), static_cast<uint32_t>(elevation), static_cast<uint32_t>(radius), oid);
-    mapFile.map_scripts[section].push_back(script);
-    mapFile.scripts_in_section[section] = static_cast<int>(mapFile.map_scripts[section].size());
-
-    return recordScriptEdit("Add Spatial Script", section, nullptr, std::move(beforeSection), 0, -1);
+    return _scriptService.addSpatialScript(programIndex, tile, elevation, radius);
 }
 
 bool ObjectCommandController::pushCommand(UndoCommand cmd) {
